Create StrechyWords
Sometimes people repeat letters to represent extra feeling, such as "hello" -> "heeellooo", "hi" -> "hiiii". In these strings like "heeellooo", we have groups of adjacent letters that are all the same: "h", "eee", "ll", "ooo".For some given string S, a query word is stretchy if it can be made to be equal to S by any number of applications of the following extension operation.
Giving a list of query words, returning the number of words that are stretchy.